// src/app/shared/models/tracker.model.ts
export interface TrackerPVModel {
    IdPVTracker: number;
    Latitude: number;
    Longitude: number;
    RazonSocial: string;
    Nit: string;
    NProblacion: string;
    SectorEco: string;
    Direccion: string;
    Representante: string;
    Contacto: string;
    TelefContacto: string;
    CorreoContacto: string;
    CargoContacto: string;
    IdUsuario: string;
    IdCampani: string;
}